Create a final tibble holding parameters for cluster markers heatmaps generation.

Compared to the cluster_markers_heatmap_params target, the output tibble is expanded to each marker test type etc.

Usage

cluster_markers_heatmaps_df_fn(
  cluster_markers_processed,
  cluster_markers_heatmap_params,
  out_dir
)

Arguments

cluster_markers_processed

(input target) A tibble.

cluster_markers_heatmap_params

(input target) A tibble.

out_dir

A character scalar: path to output directory for heatmap PDFs.

Value

A tibble. Output target: cluster_markers_heatmaps_df