scdrake: A Pipeline For 10x Chromium Single-Cell RNA-seq Data Analysis

This pipeline is implemented in the drake Make-like toolkit and provides the following steps: input quality control, filtering of cells/genes, normalization, clustering, cluster markers, differential expression analysis (constrasts -> cluster vs. cluster), integration of multiple samples. All main parts are accompanied by rich graphical outputs and HTML reports (RMarkdown).

Package options

See get_scdrake_default_options().

Included datasets

For pipeline testing two datasets from 10x Genomics are included:

PBMC 1k (v3 chemistry, Cell Ranger 3.0.0)

PBMC 3k (v1 chemistry, Cell Ranger 1.1.0). This dataset is used for testing of integration pipeline.

All credits for these datasets go to 10x Genomics. Visit https://www.10xgenomics.com/resources/datasets for more information.

Directories with dataset files can be retrieved by system.file("extdata", "pbmc1k", package = "scdrake", mustWork = TRUE) (or "pbmc3k", respectively).

See also

Author

Maintainer: Jiri Novotny jiri.novotny@img.cas.cz (ORCID)

Authors:

Jan Kubovciak jan.kubovciak@img.cas.cz