scdrake: A pipeline for droplet-based single-cell RNA-seq data secondary analysis implemented in the drake Make-like toolkit for R language

This pipeline is implemented in the drake Make-like toolkit (https://github.com/ropensci/drake) and provides the following steps of scRNA-seq secondary analysis: input quality control, filtering of cells/genes, normalization, clustering, cell type annotation, cluster markers, differential expression analysis (constrasts -> cluster vs. cluster), integration of multiple samples. All main parts are accompanied by rich graphical outputs and HTML reports (RMarkdown). Thanks to the drake package, the pipeline is highly reproducible, scalable and efficient, provides easy access to all intermediate results, and can be arbitrarily extended as pipeline definitions are abstracted as R objects. A command line interface is available to perform the most important actions directly in shell terminal.

Package options

See get_scdrake_default_options().

Included datasets

For pipeline testing two datasets from 10x Genomics can be automatically downloaded:

download_pbmc1k(): PBMC 1k (v3 chemistry, Cell Ranger 3.0.0)

download_pbmc3k(): PBMC 3k (v1 chemistry, Cell Ranger 1.1.0). This dataset is used for testing of integration pipeline.

All credits for these datasets go to 10x Genomics. Visit https://www.10xgenomics.com/resources/datasets for more information.

See also

Author

Maintainer: Jiri Novotny jiri.novotny@img.cas.cz (ORCID)

Authors:

Jan Kubovciak jan.kubovciak@img.cas.cz

Other contributors:

Michal Kolar kolarmi@img.cas.cz [funder]